# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: KeyControl API
  version: "1.0.0"
  description: |
    # Overview
    Welcome to the official API documentation for the KeyControl gateway platform. This API provides the programmatic interface for securely managing, routing, and validating your API keys.

    ## Authentication
    This API is secured using Bearer authentication. With the exception of public authentication routes (`/api/auth/*`), all endpoints require a valid token passed in the `Authorization` header:

    `Authorization: Bearer <YOUR_TOKEN>`

    **Supported Token Types:**
    * **JWT:** Obtained via the standard user login flow.
    * **Master API Key:** A system-level key (prefixed with `mk-`) used for administrative access and automated integrations.
servers:
  - url: "{scheme}://{host}:{port}/api"
    description: Admin API
    variables:
      scheme:
        default: http
        enum: [http, https]
      host:
        default: localhost
      port:
        default: "3000"
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or Master API Key
      description: |
        Pass either:
        • A JWT returned by `POST /auth/login` (or `/auth/login/verify-2fa`).
        • The Master API Key (prefix `mk-`) generated via the Organization endpoints.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
            - METHOD_NOT_ALLOWED
            - PAYLOAD_TOO_LARGE
            - BAD_GATEWAY
            - GATEWAY_TIMEOUT
        details:
          description: Optional field-level validation details
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
...
paths:
...
  /auth/login/verify-2fa:
    post:
      tags: [Auth]
      operationId: verify2FA
      summary: Complete login with 2FA TOTP code
      description: Re-validates the ADMIN_TOKEN and verifies the TOTP code. Returns a JWT on success.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, totp_code]
              properties:
                token:
                  type: string
                  minLength: 1
                totp_code:
                  type: string
                  minLength: 6
                  maxLength: 6
                  description: 6-digit TOTP code from authenticator app
      responses:
        "200":
          description: 2FA verified — JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
        "400":
          description: Bad request (e.g. 2FA not enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid admin token or 2FA code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Organization ───────────────────────────────────────────────────
...
  /organization/organization-code:
    put:
      tags: [Organization]
      operationId: updateOrganizationCode
      summary: Update organization code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organization_code]
              properties:
                organization_code:
                  type: string
                  pattern: "^[a-z0-9]{6}$"
                  description: Exactly 6 lowercase alphanumeric characters
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  organization_code:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organization/two-factor/generate:
    post:
      tags: [Organization]
      operationId: generate2FA
      summary: Generate a new 2FA secret and QR code
      responses:
        "200":
          description: 2FA secret generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  qr_code:
                    type: string
                    description: Base64-encoded data-URI of the QR code image
                  otpauth_url:
                    type: string
                  manual_entry_key:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /organization/two-factor/disable:
    post:
      tags: [Organization]
      operationId: disable2FA
      summary: Disable 2FA (requires current TOTP code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          description: 2FA not enabled or invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid TOTP code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organization/session-timeout:
    put:
      tags: [Organization]
      operationId: updateSessionTimeout
      summary: Update JWT session timeout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_timeout_seconds]
              properties:
                session_timeout_seconds:
                  type: integer
                  minimum: 120
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  session_timeout_seconds:
                    type: integer
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organization/ip-logging:
    put:
      tags: [Organization]
      operationId: updateIpLogging
      summary: Toggle IP address logging for gateway requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [log_ip_addresses]
              properties:
                log_ip_addresses:
                  type: boolean
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  log_ip_addresses:
                    type: boolean
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organization/master-key/generate:
    post:
      tags: [Organization]
      operationId: generateMasterKey
      summary: Generate or regenerate the Master API Key
      description: |
        Only JWT-authenticated admins may call this endpoint (not via Master Key auth).
        Returns the plaintext key **once** — it cannot be retrieved again.
      responses:
        "200":
          description: Master key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  master_api_key:
                    type: string
                    description: Plaintext key — save it now; it will not be shown again
                  prefix:
                    type: string
                  message:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — cannot manage master key via master key auth
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /organization/master-key:
    delete:
      tags: [Organization]
      operationId: revokeMasterKey
      summary: Revoke the Master API Key
      description: Only JWT-authenticated admins may call this endpoint.
      responses:
        "200":
          description: Master key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Resources ─────────────────────────────────────────────────────────
...